switched nodesize from int to double

+++ src/biologicalObjects/nodes/BiologicalNodeAbstract.java
@@ -17,7 +17,7 @@ public abstract class BiologicalNodeAbstract extends GraphElementAbstract {
 	private DAWISNode dawisNode;
 	private String compartment = Elementdeclerations.cytoplasma;
 	//private HashMap<Integer, Integer> values = new HashMap<Integer, Integer>();
-	private int nodesize = 1;
+	private double nodesize = 1;
 	private Vector<CollectorNode> collectorNodes = new Vector<CollectorNode>();
 	private BiologicalNodeAbstract parentNode;
 	// contains information on microarray data for this element
@@ -218,11 +218,11 @@ public abstract class BiologicalNodeAbstract extends GraphElementAbstract {
 		//setShape(vs.getEllipse(getVertex()));
 	}
 
-	public int getNodesize() {
+	public double getNodesize() {
 		return nodesize;
 	}
 
-	public void setNodesize(int nodesize) {
+	public void setNodesize(double nodesize) {
 		this.nodesize = nodesize;
 	}
